HydroMech.jl API

HydroMech.HydroMech — Module

Main module for HydroMech.jl

A collection of hydro-mechanical solvers for incompressible and compressible 2-phase flow.

HydroMech.AbstractFlow — Type

AbstractFlow{NDIMS, NVARS} An abstract supertype of specific equations such as the compressible Euler equations. The type parameters encode the number of spatial dimensions (nDim) and the number of primary variables (nVar) of the physics model.

NOTE: formulation borrowed from Trixi.jl, changed from AbstractEquations

HydroMech.CompressibleTPF — Method

CompressibleTPF

The compressible two-phase flow equations without the inertial term.

i). Total momentum (solid matrix and fluid)

\[\nabla \cdot \underline{\underline{\sigma}} + g \rho^{[t]} = 0\]

ii). Fluid momentum

\[v^{[D]} = -\frac{k^{[\phi]}}{\eta^{[f]}}(\nabla p^{[f]}-\rho^{[f]}g)\]

iii). Compressible solid mass

\[\nabla \cdot v^{[s]} = -\frac{1}{K^{[d]}}(\frac{D^{[s]} p^{[t]}}{D t} - \alpha \frac{D^{[f]} p^{[f]}}{Dt}) - \frac{p^{[t]}-p^{[f]}}{\eta^{[\phi]}(1-\phi)}\]

iv). Compressible fluid mass

\[\nabla \cdot v^{[D]} = \frac{\alpha}{K^{[d]}}(\frac{D^{[s]} p^{[t]}}{D t} - \frac{1}{B} \frac{D^{[f]} p^{[f]}}{Dt}) + \frac{p^{[t]}-p^{[f]}}{\eta^{[\phi]}(1-\phi)}\]

HydroMech.IncompressibleTPF — Method

IncompressibleHydroMech

The incompressible two-phase flow equations without the inertial term.

i). Total momentum (solid matrix and fluid)

\[\nabla \cdot \underline{\underline{\sigma}} + g \rho^{[t]} = 0\]

ii). Fluid momentum

\[v^{[D]} = -\frac{k^{[\phi]}}{\mu^{[f]}}(\nabla p^{[f]}-\rho^{[f]}g)\]

iii). Incompressible solid mass

\[\nabla \cdot v^{[s]} = - \frac{p^{[t]}-p^{[f]}}{\eta^{[\phi]}(1-\phi)}\]

iv). Incompressible fluid mass

\[\nabla \cdot v^{[D]} = \frac{p^{[t]}-p^{[f]}}{\eta^{[\phi]}(1-\phi)}\]